+++ testsuite/src/test/java/org/ironjacamar/common/metadata/ds/DataSources11TestCase.java
@@ -256,7 +256,7 @@ public class DataSources11TestCase
       Timeout t = d.getTimeout();
       assertNotNull(t);
       assertEquals(20000L, (long)t.getBlockingTimeoutMillis());
-      assertEquals(4L, (long)t.getIdleTimeoutMinutes());
+      assertEquals(4, (int)t.getIdleTimeoutMinutes());
       assertEquals(120L, (long)t.getQueryTimeout());
       assertEquals(100L, (long)t.getUseTryLock());
       assertEquals(2L, (long)t.getAllocationRetry());
@@ -354,7 +354,7 @@ public class DataSources11TestCase
       t = xd.getTimeout();
       assertNotNull(t);
       assertEquals(20000L, (long)t.getBlockingTimeoutMillis());
-      assertEquals(4L, (long)t.getIdleTimeoutMinutes());
+      assertEquals(4, (int)t.getIdleTimeoutMinutes());
       assertEquals(120L, (long)t.getQueryTimeout());
       assertEquals(100L, (long)t.getUseTryLock());
       assertEquals(2L, (long)t.getAllocationRetry());
