+++ testsuite/src/test/java/org/ironjacamar/common/metadata/ds/DataSources12TestCase.java
@@ -272,7 +272,7 @@ public class DataSources12TestCase
       Timeout t = d.getTimeout();
       assertNotNull(t);
       assertEquals(20000L, (long)t.getBlockingTimeoutMillis());
-      assertEquals(4L, (long)t.getIdleTimeoutMinutes());
+      assertEquals(4, (int)t.getIdleTimeoutMinutes());
       assertEquals(120L, (long)t.getQueryTimeout());
       assertEquals(100L, (long)t.getUseTryLock());
       assertEquals(2L, (long)t.getAllocationRetry());
@@ -385,7 +385,7 @@ public class DataSources12TestCase
       t = xd.getTimeout();
       assertNotNull(t);
       assertEquals(20000L, (long)t.getBlockingTimeoutMillis());
-      assertEquals(4L, (long)t.getIdleTimeoutMinutes());
+      assertEquals(4, (int)t.getIdleTimeoutMinutes());
       assertEquals(120L, (long)t.getQueryTimeout());
       assertEquals(100L, (long)t.getUseTryLock());
       assertEquals(2L, (long)t.getAllocationRetry());
