+++ testsuite/src/test/java/org/ironjacamar/common/metadata/ds/DataSources13TestCase.java
@@ -274,7 +274,7 @@ public class DataSources13TestCase
       Timeout t = d.getTimeout();
       assertNotNull(t);
       assertEquals(20000L, (long)t.getBlockingTimeoutMillis());
-      assertEquals(4L, (long)t.getIdleTimeoutMinutes());
+      assertEquals(4, (int)t.getIdleTimeoutMinutes());
       assertEquals(120L, (long)t.getQueryTimeout());
       assertEquals(100L, (long)t.getUseTryLock());
       assertEquals(2L, (long)t.getAllocationRetry());
@@ -389,7 +389,7 @@ public class DataSources13TestCase
       t = xd.getTimeout();
       assertNotNull(t);
       assertEquals(20000L, (long)t.getBlockingTimeoutMillis());
-      assertEquals(4L, (long)t.getIdleTimeoutMinutes());
+      assertEquals(4, (int)t.getIdleTimeoutMinutes());
       assertEquals(120L, (long)t.getQueryTimeout());
       assertEquals(100L, (long)t.getUseTryLock());
       assertEquals(2L, (long)t.getAllocationRetry());
